move role.go to separate package
